# docs/server/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Login:
    properties:
      login:
        type: string
      meta:
        items:
          $ref: '#/definitions/entity.Meta'
        type: array
      name:
        type: string
      password:
        type: string
      uri:
        type: string
    type: object
  entity.Meta:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: dima.derkach2004@gmail.com
    name: Derkach Dmitriy
    url: https://github.com/LorezV
  description: Gophkeeper project
  title: Gophkeeper Server
  version: 1.0.0
paths:
...
  /user/logins:
    get:
      consumes:
      - application/json
      description: fetching user logins
      operationId: get_logins
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Login'
            type: array
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get user logins
      tags:
      - Logins
    post:
      consumes:
      - application/json
      description: new user login
      operationId: add_login
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: card for save
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Login'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Add user login
      tags:
      - Logins
  /user/logins/{id}:
    delete:
      description: del user login
      operationId: del_login
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: login ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete user login
      tags:
      - Logins
    patch:
      consumes:
      - application/json
      description: update user login
      operationId: update_login
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Login ID
        in: path
        name: id
        required: true
        type: string
      - description: card for update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update user login
      tags:
      - Logins
...
swagger: "2.0"